move asserts
