+++ core/derived/mod.ts
@@ -1,8 +1,5 @@
-export * as colors from "./colors.ts"
+// moderate
 
-// moderate --exclude colors.ts
-
-export * from "./Date.ts"
 export * from "./Record.ts"
 export * from "./Tuple.ts"
 export * from "./Union.ts"
