+++ docs/NOTES.md
@@ -2,7 +2,7 @@
 
 - prioritizing correctness over speed, hence utilize structured outputs for the corrections request
 - root wrapping behavior
-- async assertions
+- assertions + async assertions
 - type visitor
 - how it works?
 - hash type
