+++ examples/character.eg.ts
@@ -1,7 +1,8 @@
 import Openai from "openai"
 import "@std/dotenv/load"
 import { ResponseFormat, T } from "structured-outputs"
-import { asserts, dbg } from "../util/mod.ts"
+import { asserts, ColorHex, Date } from "structured-outputs/std"
+import { dbg } from "../util/mod.ts"
 
 const greeting = T.taggedUnion("greeting", {
   Hi: T.string,
@@ -10,13 +11,12 @@ const greeting = T.taggedUnion("greeting", {
 })
 
 const Character = T.object({
-  name: T.string.assert(asserts.string.minLength, 4).assert(
-    asserts.string.maxLength,
-    30,
-  ),
+  name: T.string
+    .assert(asserts.string.minLength, 4)
+    .assert(asserts.string.maxLength, 30),
   home: T.string`The name of a fictional realm of magic and wonder.`,
   disposition: T.enum("Optimistic", "Reserved", "Inquisitive"),
-  born: T.Date`Date the character was born. Make sure it aligns with the age.`,
+  born: Date`Date the character was born. Make sure it aligns with the age.`,
   stateOfAffairs: T.Tuple(
     T.string`Home life.`,
     T.string`Professional life.`,
@@ -25,7 +25,7 @@ const Character = T.object({
   randomValue: T.Union(T.string, T.number),
   friends: T.array(T.string)`Names of the character's friends.`,
   greeting,
-  favoriteColor: T.colors.ColorHex,
+  favoriteColor: ColorHex,
 })
 
 const response_format = ResponseFormat("create_character", Character)`
