+++ mod.ts
@@ -2,3 +2,4 @@
 
 export * from "./client/mod.ts"
 export * from "./core/mod.ts"
+export * from "./std/mod.ts"
