+++ std/mod.ts
@@ -1,0 +1,6 @@
+export * as asserts from "./asserts/mod.ts"
+
+// moderate --exclude asserts
+
+export * from "./colors.ts"
+export * from "./temporal.ts"
