+++ std/temporal.ts
@@ -1,17 +1,17 @@
-import { assert, asserts } from "../../util/mod.ts"
-import type { Type } from "../Type.ts"
-import { number, object, transform } from "../types.ts"
+import { T, type Type } from "../mod.ts"
+import { assert } from "../util/mod.ts"
+import { max, min } from "./asserts/number.ts"
 
-const ZeroBasedInteger = number.assert(asserts.number.min, 0)`Zero based.`
+const ZeroBasedInteger = T.number.assert(min, 0)`Zero based.`
 
 const YearMonthDay: Type<{
   year: number
   month: number
   day: number
-}> = object({
-  year: number,
-  month: ZeroBasedInteger.assert(asserts.number.max, 11),
-  day: number.assert(asserts.number.min, 1).assert(asserts.number.max, 31),
+}> = T.object({
+  year: T.number,
+  month: ZeroBasedInteger.assert(max, 11),
+  day: T.number.assert(min, 1).assert(max, 31),
 })`Ensure the day is valid for corresponding year and month.`
   .assert(({ year, month, day }) => {
     const date = new Date(year, month, day)
@@ -23,16 +23,14 @@ const YearMonthDay: Type<{
   })
 
 export { Date_ as Date }
-const Date_: Type<Date> = transform(
+const Date_: Type<Date> = T.transform(
   "Date",
-  object({
+  T.object({
     yearMonthDay: YearMonthDay,
-    hour: ZeroBasedInteger.assert(asserts.number.max, 23),
-    minute: ZeroBasedInteger.assert(asserts.number.max, 59),
-    second: ZeroBasedInteger.assert(asserts.number.max, 59),
-    millisecond: ZeroBasedInteger
-      .assert(asserts.number.max, 0)
-      .assert(asserts.number.max, 999),
+    hour: ZeroBasedInteger.assert(max, 23),
+    minute: ZeroBasedInteger.assert(max, 59),
+    second: ZeroBasedInteger.assert(max, 59),
+    millisecond: ZeroBasedInteger.assert(max, 0).assert(max, 999),
   }),
   ({ yearMonthDay: { year, month, day }, hour, minute, second, millisecond }) =>
     new Date(year, month, day, hour, minute, second, millisecond),
